drawing pixels and rec now possible!
The ability to draw colored pixels "safely" is not implemented. This involves getting a `framebuffer` from the bootloader and passing it to the custom framebuffer (`FBuffer`) struct that has impl functions that have the ability to draw rectangles and pixels depending on if the correct information is provided.

### Example Implementation
---
```rust
let buffer = FBuffer::new(&boot_info.framebuffer);

// Lets draw a rectangle at pos x = 0, and y = 0 with a size of x + 10, y + 10
//                pos     size     COLOR
//              (x, y), (x+, y+)  HEX CODE
buffer.draw_rec((0, 0), (10, 10), 0xFFFFFF);
```