over-stacked: Added documentation, and started on HeaplessMap
HeaplessMap is a HashMap like type that contains a key, and a value. This will allow the caller to use keys instead of indexs just like in a python dict. This works by storing two sepreate `HeaplessVec`'s and using one for the Keys, and one for the Values. As long as your Key type has `==`, then we can use it to find your Value.