Alloc: Fixed inplace change of Free to Used
We had an issue where we would just replace the old entry with `used`. This, however, was not the best idea because we rely on the ptr being aligned. This means that we were giving out a ptr that was aligned, but never stored its alignment. This caused us to have issues when freeing because we would have a mismatch and would be unable to free.