Alloc: Added String lib
The string library was added to the allocation library. This library includes implementation of a basic string for heap allocation. The String struct is defined here along with its functionalities like push (char,into the string), push_str (a string into another string), and as_str (return string form of the data). It allows for comparison of strings and allows strings to be created from different data types. The implementation also includes deref and write implementations. The tests for the String struct are also created ensuring its smooth operation.